数据库添加摘要等字段，增加文件上传和编辑器功能

+++ src/main/java/cn/yx/service/ActivitiesService.java
@@ -54,7 +54,7 @@ public class ActivitiesService {
         }
         activities.setBrowses(browses);
         activities.setUrl(url);
-        activities.setStatus(status);
+        activities.setStatus(status.byteValue());
         whsActivitiesMapper.insertSelective(activities);
         return activities;
     }
